// src/api/billard/generated/models/GameStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Aktueller Status des Spiels
 */
export enum GameStatus {
    NOT_STARTED = 'NOT_STARTED',
    IN_PROGRESS = 'IN_PROGRESS',
    PAUSED = 'PAUSED',
    FINISHED = 'FINISHED',
    ABORTED = 'ABORTED',
}